HP Progress Redo
Redid my work from the Early Access in the Main Branch to avoid complications from merging. Fixed an issue with damage so now damage should correctly vary in accordance with the weapons damage value instead of defaulting to 1. The Dummies have varying HP, with the one on the left having 3, the center one 6, and the right one 9